Add buffer to network_usage
Added a basic buffer to the network_usage. This just smoothes out the
traffic data by a bit. The default buffer is configured with the size 5
and works pretty well.

Also removed the `license-file` from the `Cargo.toml`. I made this
change in the RustFmt and Clippy fix PR already but it seems like I
accidentally reverted it. This just removes a compiler warning.