Replace LoopFN with interval_at
After dogamak/xcbars#44 already removes LoopFN from network_usage, this
also removes it from `pipe.rs`. As a result it has also been deleted
from the `utils` module and the `stream_loop_fn.rs` file itself has been
deleted.

This is not a stable commit unless dogamak/xcbars#44 is added. Once that
is done I'll rebase this on top of it and it should be ready to merge.

This fixes #45.